requirements.txt
python-telegram-bot==20.7
requests==2.32.4
python-dotenv==1.0.0
psycopg[binary]==3.3.3
Flask==2.3.3
